requirements.txt
asgiref
Django
thefuzz
gunicorn
isodate
pyparsing
rdflib
six
SPARQLWrapper
sqlparse
tzdata
whitenoise
django-environ
elasticsearch
elasticsearch-dsl
django-elasticsearch-dsl
django-countries
django-markdownify